Feature/quillconnect/move questions to lms (#5570)
* WIP fixing typescript errors

* Fix typescript errors

These appeared in our existing code when we converted the file from js to ts

* Update Question-related API payloads

These need to match the expected shape (data inside a specifically-shaped key) since wraps_parameters won't allow wild-carding

* Fix some weird history with accidentally committed stuff

* Remove Firebase from question actions

* Tweak update_all routes based on tweak to backend

* Reference question UID properly when reloading newly created items

* Add some type definitions for the API

* Tweak properties and update comment
