+++ services/QuillConnect/app/components/itemLevels/itemLevel.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import {connect} from 'react-redux'
 import ItemLevelForm from './itemLevelForm.jsx'
 import levelActions from '../../actions/item-levels.js'
-import questionActions from '../../actions/questions.js'
+import questionActions from '../../actions/questions'
 import _ from 'lodash'
 
 const ItemLevel = React.createClass({
