+++ services/QuillConnect/app/constants.js
@@ -55,6 +55,7 @@ export default {
 
   // QUESTION ACTIONS
   RECEIVE_QUESTIONS_DATA: 'RECEIVE_QUESTIONS_DATA',
+  RECEIVE_QUESTION_DATA: 'RECEIVE_QUESTION_DATA',
   AWAIT_NEW_QUESTION_RESPONSE: 'AWAIT_NEW_QUESTION_RESPONSE',
   RECEIVE_NEW_QUESTION_RESPONSE: 'RECEIVE_NEW_QUESTION_RESPONSE',
   START_QUESTION_EDIT: 'START_QUESTION_EDIT',
