+++ services/QuillConnect/newSess.tmp
@@ -1,0 +1,1 @@
+{"answeredQuestions":[{"question":"-KRjGXoreoJZsAUqiccS","attempts":[{"response":{"child_count":202,"conceptResults":{"GiUZ6KPkH958AT8S413nJg":{"conceptUID":"GiUZ6KPkH958AT8S413nJg","correct":false,"key":"GiUZ6KPkH958AT8S413nJg"}},"concept_results":[{"conceptUID":"GiUZ6KPkH958AT8S413nJg","correct":false}],"count":20909,"feedback":"<p>Revise your work. Add to the second part of your sentence to make it a complete thought.</p>","first_attempt_count":9106,"id":12299,"key":"12299","optimal":false,"question_uid":"-KRjGXoreoJZsAUqiccS","text":"Owls live in trees, or underground."}},{"response":{"child_count":113317,"conceptResults":{"GiUZ6KPkH958AT8S413nJg":{"conceptUID":"GiUZ6KPkH958AT8S413nJg","correct":true},"UOnnSsgGF3iMAMFlHdfcdw":{"conceptUID":"UOnnSsgGF3iMAMFlHdfcdw","correct":true}},"concept_results":[{"conceptUID":"GiUZ6KPkH958AT8S413nJg","correct":false}],"count":224541,"feedback":"<p>That&#x27;s a strong sentence!</p>","first_attempt_count":117547,"id":12262,"key":"12262","optimal":true,"question_uid":"-KRjGXoreoJZsAUqiccS","text":"Owls live in trees, or they live underground."}}]}],"currentQuestion":"-KRjGpCS0R4g6Nnkv5tX","questionSet":["-KRjGXoreoJZsAUqiccS","-KRjGpCS0R4g6Nnkv5tX","-KRjHBnLs-6GxgqITD0S","-KRjH47ZxG6XQxQ97PIV","-KRjGxHJSolOcT4uhgkb","-KRjGf869g3TJdrvYUSW"],"unansweredQuestions":["-KRjHBnLs-6GxgqITD0S","-KRjH47ZxG6XQxQ97PIV","-KRjGxHJSolOcT4uhgkb","-KRjGf869g3TJdrvYUSW"]}
